+++ index.html
@@ -10,5 +10,6 @@
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
     <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
+    <script src="https://static.nid.naver.com/js/naveridlogin_js_sdk_2.0.2.js"></script>
   </body>
 </html>
